requirements.txt
pytest
pyyaml
allure-pytest
colorlog
playwright
pytest-playwright
colorama
